server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
python-dotenv==1.0.1
pyautogui==0.9.54
pydantic==2.9.2
